[java] Update Tests for JEP 463: Implicitly Declared Classes and Instance Main Methods
