+++ pmd-java/etc/grammar/Java.jjt
@@ -1,4 +1,6 @@
 /**
+ * Support "JEP 463: Implicitly Declared Classes and Instance Main Methods (Second Preview)" (Java 22)
+ *     No changes.
  * Support "JEP 459: String Templates (Second Preview)" (Java 22)
  *     Use ASTTemplate.setContent instead of setImage.
  * Remove support for Record pattern in enhanced for statements. This was only a Java 20 Preview feature.
@@ -1153,7 +1155,7 @@ ASTCompilationUnit CompilationUnit() :
   // OrdinaryCompilationUnit: -> TopLevelClassOrInterfaceDeclaration
   ( TypeDeclaration() ( EmptyDeclaration() )* )*
 
-  // UnnamedClassCompilationUnit:
+  // SimpleCompilationUnit:
   [
       ( LOOKAHEAD(3) ClassMemberDeclarationNoMethod() )*
       ModifierList() MethodDeclaration()
